# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Clientes e Lotes",
        "description": "Documentação da API (Laravel 11).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor"
        }
    ],
    "paths": {
...
        "/api/lotes": {
            "get": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Listar lotes",
                "operationId": "0bb575f5c36032c445a3ca5ee9131c79",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 120,
                            "minLength": 2
                        }
                    },
                    {
                        "name": "num_loteamento",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "num_quadra",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "num_lote",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "area_lote",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "minimum": 0
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "nome",
                                "num_loteamento",
                                "num_quadra",
                                "num_lote",
                                "area_lote",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "dir",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LotePage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Filtro inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Criar lote",
                "operationId": "e8facbb615018d87d8a7fb9e0842e751",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoteStore"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lote"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos mal formatados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito de localização",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Regra de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lotes/{id}": {
            "get": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Detalhar lote",
                "operationId": "b5d35a1061c5511e387a1b1088288029",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lote"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Atualizar lote (total)",
                "operationId": "54c5500056ebc432f7d61280e691ef6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lote"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos mal formatados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito de localização",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Regra de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Excluir lote",
                "operationId": "a5e5e3d0e53c34dbb503eee4c2db6b50",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excluído",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Atualizar lote (parcial)",
                "operationId": "5dad7576ac736c761f75c64c9f64adda",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lote"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos mal formatados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito de localização",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Regra de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Erro de validação"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "schema": "CommonSchemas"
                        }
                    }
                },
                "type": "object"
            },
...
            "Lote": {
                "required": [
                    "id",
                    "nome",
                    "num_loteamento",
                    "num_quadra",
                    "num_lote"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "num_loteamento": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "num_quadra": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "num_lote": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "cliente_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "area_lote": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "LoteStore": {
                "required": [
                    "nome",
                    "num_loteamento",
                    "num_quadra",
                    "num_lote"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "num_loteamento": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "num_quadra": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "num_lote": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "cliente_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "area_lote": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "LoteUpdate": {
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "num_loteamento": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "num_quadra": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "num_lote": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "cliente_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "area_lote": {
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "LotePage": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Lote"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            }
        }
    },
...
}